+++ eslint.config.js
@@ -1,13 +1,15 @@
-import { js, prettier } from '@studiometa/eslint-config';
-import { globals } from '@studiometa/eslint-config/utils';
+import { defineConfig, js, prettier, globals } from '@studiometa/eslint-config';
 
-export default [
-  ...js,
-  ...prettier,
+export default defineConfig(
+  js,
+  prettier,
   {
     files: ['packages/preset-*/**/*', 'packages/webpack-config/**/*'],
     languageOptions: {
       globals: globals.node,
     },
   },
-];
+  {
+    ignores: ['**/dist/**'],
+  },
+);