+++ packages/demo/src/templates/pages/posts/bar.js
@@ -1,7 +1,8 @@
 import { data as defaultData } from './[slug].js';
 
 /**
- *
+ * @param {{ slug: string }} props
+ * @returns {Promise<Record<string, unknown>>}
  */
 export async function data({ slug } = {}) {
   const mainData = await defaultData({ slug });